requirements.txt
# ============================================
# BookRank - 纽约时报畅销书排行榜
# Python 依赖包 (Render 免费版优化版)
# 删除未使用依赖减少内存和启动时间
# ============================================

# 核心框架
Flask==3.1.3
Werkzeug==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.1
Flask-Caching==2.1.0
Flask-Talisman==1.1.0
Flask-Limiter==3.10.1
Flask-WTF==1.2.2

# ORM 和数据库
WTForms==3.2.1
SQLAlchemy==2.0.38
alembic==1.14.0
psycopg2-binary==2.9.10

# 配置和环境
python-dotenv==1.0.1

# HTTP 客户端（只保留 requests，httpx 未使用）
requests==2.32.3

# WSGI 服务器（Render 生产环境需要）
gunicorn==23.0.0

# HTML 解析（beautifulsoup4 自带 html.parser，去掉 lxml）
beautifulsoup4==4.13.0

# AI 翻译（智谱AI主用，deep-translator备用回退）
zhipuai>=2.0.0
deep-translator>=1.11.0

# 工具库（去掉 cachetools 未使用，保留 tenacity 重试）
tenacity==9.0.0

# 任务调度（内存队列，免费版 Render 无 Redis 所以不用 Celery）
APScheduler==3.11.0

# 导出功能（PDF/Excel）
fpdf2==2.7.9
openpyxl==3.1.2
Pillow==11.1.0

# 监控
sentry-sdk==2.22.0
